requirement.txt
fastapi
uvicorn[standard]
gitpython
python-dotenv
requests
openai
pydantic